// src/shared/api/generated/models/integrationActivityListParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { IntegrationActivityListProviderItem } from "./integrationActivityListProviderItem";
import type { IntegrationActivityListActivityTypeItem } from "./integrationActivityListActivityTypeItem";
import type { IntegrationActivityListStatusItem } from "./integrationActivityListStatusItem";

export type IntegrationActivityListParams = {
  /**
   * @maxLength 2000
   */
  cursor?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minItems 1
   * @maxItems 10
   */
  provider?: IntegrationActivityListProviderItem[];
  /**
   * @minItems 1
   * @maxItems 2
   */
  activityType?: IntegrationActivityListActivityTypeItem[];
  /**
   * @minItems 1
   * @maxItems 5
   */
  status?: IntegrationActivityListStatusItem[];
  /**
   * @minLength 1
   * @maxLength 200
   */
  externalUserId?: string;
  createdFrom?: string;
  createdTo?: string;
};